Generate bookmarks for all git refs
Summary:
Add a flag to `gitimport`: `--generate-bookmarks` that can be used with `full-repo` to
* Generate a bonsai changeset for all git commits
* Generate a bonsai bookmark for all git refs

This means that all bonsai changesets will be public and derived data can conveniently be backfilled at once.

Note: this should not be run on a `mononoke` repo that is already in production (with things like the warm bookmark cache enabled) as that would generate a large load on our infrastructure.

## Intended use:

### On importing a git repo,
* Create a mononoke repo and use `gitimport --generate-bookmarks full-repo` to import all changesets and bookmarks in the mononoke repo.
* Backfill all derived data types on the full repo.
* Enable the repo in production.

### On pushes to the repo (with hooks)
* Use `remote_gitimport` to import all new commits that were pushed.
* Use `git_move_bookmark` to move any updated bookmark to their new location.

Reviewed By: mitrandir77

Differential Revision: D41534200

fbshipit-source-id: 2ffd33807da7d978fdee959112107d0c7a2d326c