Add emacs style navigation keybindings to crecord
Summary:
This diff introduces emacs style bindings for navigating the crecord interface for people who are used to using magit.
C-n: Move down
C-p: Move up
C-v: Move to the next item of the current item type you have selected
M-v: Move to the previous item of the current item type you have selected
tab: Fold / Unfold item

curses.raw() is used so that C-v does not need to be pressed twice for it to work.

Reviewed By: quark-zju

Differential Revision: D41539621

fbshipit-source-id: 873c605c793d8ceba4d3f630aaca617506a19e43