add Troubleshooting section for GitHub page
I just ran into an issue where `sl pr` failed to run because I
did not have `gh` configured as my Git credential helper.
It took me a bit of time to figure this out, so I thought I would
add it to the website, though the docs related to GitHub
integration are somewhat spread out, so I'm not sure whether this
is the best place for it. A dedicated Troubleshooting page might
be more appropriate, though such a page should probably have at
least two entries?

Though perhaps an even better solution is updating the code to
recognize this sort of failure and then suggest `gh auth setup-git`
directly when it happens.

Test Plan:

http://localhost:3000/docs/git/intro