add versioning info to privhelper
Summary:
In some cases, we may want a developer-instance of EdenFS to communicate with a production privhelper daemon. However, this can cause issues if the protocol between the edenfs-daemon and privhelper-daemon change in the developer version of EdenFS (or vice versa).

To ensure protocol versioning issues aren't a problem, we need to add versioning info/logic to the privhelper server/conn code. This will ensure that we can keep the protocols backwards/forwards compatible in the future.

Reviewed By: chadaustin

Differential Revision: D40960617

fbshipit-source-id: e862f041a3d31efd68dae6a52146e68742b1405b