prepare EdenFS tests for a world without HgImporter
Summary:
EdenFS's tests still fall back on a debugedenimporthelper process to
import trees and blobs. That's because the test repos still use the
deprecated revlog format. Migrate the test repos to use indexedlog
storage so that SaplingNativeBackingStore hits.

Note: This repo configuration is awkward and temporary, and the
Sapling team is planning to migrate to "Eager" repos to support the
`hg init` case in the future.

Reviewed By: quark-zju

Differential Revision: D41447075

fbshipit-source-id: 23c2f13dfc9f964d9386d7f4c920222db8f25144