Add prefix for network access void
Summary:
It's very annoying to have to add every new test here. And it is also confusing, because your test fails for very weird reason. It has confused me before, my intern, and other people in the team.

This is not the long term solution but it at least helps with the problem when modifying big things like SCS.

Reviewed By: mitrandir77

Differential Revision: D41522350

fbshipit-source-id: 8efba13590d1efcaec4cff33f0713b46cc6946f3