drawdag: support declaring commit properties using Python script
Summary:
Previously there is no way to set user or complex (including spaces) commit
messages using drawdag. This diff introduces a way to do that for maximum
flexibility.

This is intended to be used to create repo state for documentation rendering.
Note the `{sl}` template configs do not exist in test environment yet so we
emulate a minimal version of the template in test.

Reviewed By: zzl0

Differential Revision: D41674152

fbshipit-source-id: 2bb1212da06ddd2fd93f780d6f7b291f873fa922