Decrease log level of noisy logs
Summary:
The `flushing {} in-memory IdMap entries to SQL` log line is VERY noisy and shows in every one of our jobs, sometimes it's the most common log line.

It doesn't add that much value to us, so let's make it show up less by making it trace. (Not debug because we already run most TW jobs with debug enabled)

Differential Revision: D41548859

fbshipit-source-id: a5df753e4b7ec92d6725c1cd349ccdc1ff58ddaf