Fix issues in commit_validator
Summary:
This diff fixes the biggest issues with commit_validator and should allow it to move on and stop reporting dead detector alerts.

## The fixes

The bug it fixes is this:
https://www.internalfb.com/code/fbsource/[27145be3891913fd56cf85e570d1fa2d0d0cc4ae][history][blame]/fbcode/eden/mononoke/commit_rewriting/commit_validator/src/tail.rs?lines=124

We are filtering the bookmarks to **only have bookmarks we skip**, that is, it is doing exactly the opposite of what we want. This leads to never processing any entries and getting stuck.

The second issues it fixes is: `DifferenceOfUnionsOfAncestorsNodeStream` is slow. Specially when we add "optimisations" that exclude master bookmark. When it is looking at a 5 month old commit, it would crawl master until five months ago.

What I did is start using `RangeNodeStream` when it's a forward bookmark move, which should be >99% cases, and leave the old way as a fallback in other cases. Since this is just for validation, if we continue running into problems for non-FF moves, we can just ignore them and only look at the new value of the bookmark, which should be good enough for us.

## Explanation of what is happening in prod

In prod, we see validator take a loooong time (sometimes days) to start outputting anything. This happens because: we start at an old entry id, MILLIONS of items before master. Since we have the bug that skips all bookmarks we actually want, and there's none we actually want to skip, we never update the "latest id" value, so whenever we restart validator it has to go through those MILLIONS of items again until it catches up to master, when it starts saying "no-op".

Reviewed By: mitrandir77

Differential Revision: D41550264

fbshipit-source-id: c7de6a890a60bda85798e3133e3a5eeecf90adaa