script to detect corrupt file size
Summary:
EdenFS is sometimes displaying truncated file content through the filesystem. This is because the size EdenFS is reporting is incorrect.

This adds a checker to scan the local store for corrupted file sizes. It's too slow to put in eden doctor right now, but I will use it to stress test and see if I can repro the corruption.

Reviewed By: xavierd

Differential Revision: D41012331

fbshipit-source-id: b74ce536d0ced1a01c19a92759258594ffd49e68