`eden debug blob` add options for new fetching locations
Summary:
The daemon now supports specifying different locations where data can be
fetched from.  We can support these in the eden debug blob CLI so that it is
easy to inspect which stores have a blob and what contents those stores have
for the blob.

This diff does break the API of the `eden debug blob` command (no more --load
option). But I don't spot this uses anywhere in code and is only intended for
use by the eden team, so this seems ok.

Reviewed By: xavierd

Differential Revision: D41165545

fbshipit-source-id: 7f35de59ad8073173e917f01eeae693500f6540c