server/scs_server: rename RepoShardedProcess implementors
Summary:
These names are slightly misleading.  For Mononoke server, it's not the only
thing that uses `mononoke_api`, so its process and process executor structs
should be named `MononokeServer...`.  For the source control service, the right
way to camel-case an acronym (according to the Rust naming guidelines) is as
`ScsServer...`.

Differential Revision: D41535134

fbshipit-source-id: 6dcbba1feb5208a9b45e8e0b9e66d8cf77938416