mononoke_app: add wait_until_terminated for processes without foreground services
Summary:
Add a `wait_until_terminated` method for `MononokeApp` which runs the same
termination process as `run_until_terminated`, but does not require a
foreground server.

Previously, these kinds of services used `future::pending()`, however this
means when a termination signal receives, the shutdown *always* times out, as
the server future never completes.  For SCS server we manually handled this, so
by making it a common function we can use this in other Mononoke services.

Reviewed By: yancouto

Differential Revision: D41534838

fbshipit-source-id: acd746eb8b2f25b2c1f7475cac2490d2a542f4eb