Don't fail to create already existing bookmark
Summary:
If some bookmarks already exist in mononoke, `create_bookmark` fails. In that case, use `resolve_bookmark` to get the old changeset for that bookmark and move the bookmark from its old location to the new one.

Start by attempting to create the bookmark as it makes the default usecase, where we're running this tool on a new repo where mononoke doesn't have any bookmarks set yet fast.

Reviewed By: mitrandir77

Differential Revision: D41769964

fbshipit-source-id: 52fdea009ba7ca5339f9c8d6dab7152c15c0ff86