prefer fsPath to path when working with vscode.Uri (#305)
Summary:
Stack created with [Sapling](https://sapling-scm.com). Best reviewed with [ReviewStack](https://reviewstack.dev/facebook/sapling/pull/305).
* __->__ https://github.com/facebook/sapling/issues/305

Leverage `fsPath` for path normalization so that the path of a file
is recognized as a member of a Sapling repository based on the
`RepositoryCache` checks.

It turns out that `vscode.Uri.prototype.path` is problematic on
Windows because it is not a "canonical path." For example,
for the same folder, `.path` could be reported as either:

```
"/C:/Users/micha/src/sapling/LICENSE"
```

or:

```
"/c:/Users/micha/src/sapling/LICENSE"
```

which means that basic string `===` tests for equality will not
work. However, both will report the same value for `.fsPath`:

```
"c:\\Users\\micha\\src\\sapling\\VERSION"
```

As noted on https://code.visualstudio.com/api/references/vscode-api#Uri.fsPath:

> The string representing the corresponding file system path of this Uri.
>
> Will handle UNC paths and normalize windows drive letters to lower-case. Also uses the platform specific path separator.

Because this issue arises only on Windows, the ideal fix is to
introduce a linter that prohibits the use of `.fsPath`.

Pull Request resolved: https://github.com/facebook/sapling/pull/305

Test Plan:
Built and ran the VS Code extension on Windows in both
"normal" Windows 10 and WSL. Verified that
**Open Diff View For Current File** works in both whereas
previously it did not work in non-WSL due to the lack of
path normalization.

Reviewed By: sggutier

Differential Revision: D41829879

Pulled By: bolinfest

fbshipit-source-id: d8e9f5f5129ca69135b1ef95d2db79c703c01399