service: only allow takeover if all mounts are initialized
Summary:
The most common crash on macOS is when a graceful restart is performed while a
mount is ongoing. In this case, an EDEN_BUG is raised in EdenFS leading to
EdenFS to crash and restart.

As a very easy solution, let's check prior to starting graceful restart if
there are any ongoing mounts, and if so, just abort the graceful restart. Note
that there is still a potential race if a mount occurs right after this check,
but that should be significantly more rare and unlikely to happen in practice.

Reviewed By: kmancini

Differential Revision: D41640271

fbshipit-source-id: c0fd0cc0305cbcdd941aa0e2e422706af799d5b3