mononoke_api: log new commits in land_stack
Summary:
Currently calling `repo_land_stack` in SCS server (or later on, EdenAPI) won't log the new public commits to the repo update logger.

For `move_bookmark` and `create_bookmark`, we handle this by asking `bookmarks_movement` to do the logging, but that's not possible for `pushrebase_onto`.  Fix this by implementing this flag for `pushrebase_onto`, and setting it when calling from SCS server.

Note that up until we switched over to the land_service, this would've resulted in double-logging, but now we have the separate land service, we can fix the SCS service.

Note that none of this is particularly nice at the moment.  We log from multiple places and maintaining this is clumsy.  We don't have a single "commits are created" point to hook into, so we will need to refactor commit creation out in order to resolve this.  That will have to come later on.

Reviewed By: yancouto

Differential Revision: D41620032

fbshipit-source-id: 43b2fa983054f0b4815a6c75c84f030269428b12