Handle platform/openExternal even if ISL cannot find the sl binary when starting. (#307)
Summary:
This was reported as part of https://github.com/facebook/sapling/issues/282.
Specifically, if `sl` (or `sl.exe`) is not found on the `$PATH`,
ISL shows an error box with a **See installation docs** button,
and clicking it sends a `platform/openExternal` message to the server to open an external URL.

This message was not getting processed because the handler
was only processing messages when the `currentState` of
`ServerToClientAPI` was `"repo"` (i.e., not `loading` or `error`).

Failing to find `sl` put things in the `error` state, so the
`platform/openExternal` message was ignored.
Note that the bulk of the messages are processed by
`handleIncomingMessageWithRepo()`, which reads the
`Repository` off `this.currentState`, though that is not
present when in the `loading` or `error` state.

As such, this diff adds a conservative fix, which is to
handle only messages of type `PlatformSpecificClientToServerMessages`,
as `platform.handleMessageFromClient()` can be called
without a `Repository` object.

Pull Request resolved: https://github.com/facebook/sapling/pull/307

Test Plan: Verified clicking the **See installation docs** button works now.

Reviewed By: ndmitchell

Differential Revision: D41845324

Pulled By: bolinfest

fbshipit-source-id: 4d90f4dda022c82951f71f22206d424afcfbd4c4