admin: Lock and unlock all repos in megarepo together
Summary:
On all megarepo instructions, you either lock or unlock all repos in a megarepo group together. But you do that manually one command at a time.

Let's make a single command do all of those operations for us. Hooray automation. It's still possible to force the operation to be done on a single repo with a flag.

Ideally this would all be a single query for optimal speed, but it would overly complicate our stuff, so sequential ones will do for now, it's anyway much faster than what we have today.

Suggested by mitrandir77

Reviewed By: mitrandir77

Differential Revision: D41475493

fbshipit-source-id: 901451c5d5ffd8f9a035b587ba9ee875e0b7be10