Add configurable rendezvous controller and use it on tunables one
Summary:
This is a refactoring only change. It adds ConfigurableRendezVousController, which allows customising the controller using any provided functions. We implement `TunablesRendezVousController` by using this new, more generic controller.

The new controller isn't used without tunables on this diff, but this refactor is good because I plan to start some controllers with some more hardcoded values, not reusing the same tunables.

Reviewed By: markbt

Differential Revision: D41875741

fbshipit-source-id: fb4b7bb0e31b802c8bb044840c21a17bc23ab377