create StartupStatusSubscriber
Summary:
In D38962809 I will be allowing users to inspect EdenFS startup status.

In this change I am introducing an abstraction that allows subscribing to
eden startup status. This abstraction is generic. It's not just thrift streams
that can be subscribed here, but it is designed to fit what we need for a
thrift stream subscriber:
1. ability to publish data
2. ability to complete the stream when startup has completed.

A subscriber will provide two methods:
1. publish - publish is used when there is a startup status update.
2. destructor - the object will be destroyed when startup has completed.

StartupStatusSubscriberState will keep track of all subscribers
to startup status. This will be shared with the startup logger and eden server
so that the eden server can add subscribers and the startup logger can publish
updates.

This abstraction is thread safe and gathers subscriber book keeping in one place.

Reviewed By: mshroyer

Differential Revision: D41057243

fbshipit-source-id: 0ca2b25eb51f13166557c349f01f6edf5f60cfe0