scs_server: add generated status to metadata diff format
Summary:
The generated status of a file can be:
* Fully generated: If the file is generated (annotated with either generated or partially-generated) and doesn't contain any manual sections.
* Partially generated: If the file is generated and contains some manual sections.
* Not generated: If the file is not generated.

We also keep the ranges of manual sections in the partially generated case to be used in calculating significant lines.

Note: Some files marked with generated have manual sections, and some files marked with partially-generated don't have manual sections. So in practice there's no difference between generated and partially-generated.

Reviewed By: markbt

Differential Revision: D41767452

fbshipit-source-id: a1fcdfbaff567e530a1fcc99d8fd8a9711c41e09