reserve GitHub issue number for new pull request
Summary:
Historically, I was under the impression that it was impossible
to create a pull request with a specific number. As you can see
in this diff, we had a `guess_next_pull_request_number()` helper
function that had a "pretty good" heuristic for guessing the next
pull request number, which made it likely that we could make
the number encoded in the branch name match that of the PR.
Unfortunately:

- This heuristic was not guaranteed to work 100% of the time,
  due to either races or deleted issues that could throw off the count.
- Using this heuristic required that we create the pull requests
  serially [from bottom to top of the stack] rather than in parallel.

However, today while reading the REST API docs for creating a
pull request (https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#create-a-pull-request),
I noticed it was possible to specify an `issue`:

> An issue in the repository to convert to a pull request. The issue title, body, and comments will become the title, body, and comments on the new pull request. Required unless `title` is specified.

Ah ha! This meant we could first create a number of GitHub
issues (in parallel!) with placeholder data and then use the numbers
from these issues when creating pull requests (also in parallel)!
This also means we know the PR number when choosing the branch
name (recall the branch cannot be changed after creating the PR).

This diff updates the existing `sl pr` command to take advantage
of this information.

Reviewed By: muirdm

Differential Revision: D42050143

fbshipit-source-id: a52694f6599c4ba99f568c538d557e6d10db72a6