Log bookmarks creation
Summary:
When running gitimport on a real repo, creating bookmarks can be a slow operation. By logging it, we make it more obvious to the user that we are not stalling.

Incidentally, it helps with integration tests for improvements to the branch creation logic that come later in this stack.

Reviewed By: clara-9

Differential Revision: D42071748

fbshipit-source-id: 92313ad9f0ab2e3840c1e39cfafa08de0b43d549