Replace remotes/origin prefix with heads in refs
Summary:
One way to import a repo from git to mononoke is to clone the repo and give its path to the importer.

In that case, the remotes branch will be prepended by `refs/remotes/origin`.

In the context of the hooks which call `remote_gitimport` to keep the repo up-to-date, branches show as `refs/heads/branch_name`, so we would end up creating branches on gitimport that would only stall behind later.

Reviewed By: mitrandir77

Differential Revision: D42071750

fbshipit-source-id: 8ad28bbdf04be8c260d0dfa73fc190fd4debfbad