Only initialize necessary repos
Summary:
By default, gitimport would read every mononoke config when creating a `RepoContext` for the purpose of creating bookmarks.

This means that it would spend a lot of time loading all kinds of repos present in configerator, even though it would really only need access to one of them, the one that's passed as a CLI argument to the tool.

Only open the repo we need.

Internal
Freely inspired by T139371732 #thanks, mitrandir77

Reviewed By: clara-9

Differential Revision: D42071747

fbshipit-source-id: 129ec878de309aefdf695fb34f6a6de6e180ee94