status: fix native status to ignore modified symlinks on Windows
Summary:
Python status has some logic to ignore changes to symlinks on platforms that don't support symlinks. This diff ports that logic to Rust as well.

The issue appeared on Windows with symlinks inside node_modules/.bin directories. On Windows when you run "yarn", the symlinks are overwritten with wrapper scripts (since symlinks are not supported). The Rust status was showing those as modified files, buy Python had special logic to ignore "suspect" changes to symlinks.

Reviewed By: quark-zju

Differential Revision: D42016078

fbshipit-source-id: 18d0f41253611fab28e36aa4f720534144008d3f