config: load sapling "prod" config if TEST_PROD_CONFIGS is set
Summary:
It turns out `unsetting TESTTMP` is not a great idea for loading production
configs, since there might be other references to it, for example:

  [remotefilelog]
  cachepath=$TESTTMP/foo/bar

in test hgrc relies on runtime expansion of `$TESTTMP`.

For documentation purpose, we do want to load the "prod" config including the
templates and colors. Make it possible using `TEST_PROD_CONFIGS`.

Reviewed By: muirdm

Differential Revision: D42102951

fbshipit-source-id: baa82b85873f0793b3b873b9a63bb7865e28787b