inodes: actually ensure that a single GC can run concurrently
Summary:
The WorkingCopyGCLease destructor was slightly wrong as it always reset the
gcRunning_ flag, which is wrong as the destructor would be called on moved
objects too. Thus even if a previously obtained lease was moved into a future,
a second GC could start running as soon as the previous lease was destructed
from the stack.

Reviewed By: chadaustin

Differential Revision: D42087849

fbshipit-source-id: 8c1c8296fc24e8c4a819bd6d22ff81cad1d7b832