service: only run GC on fully mounted mounts
Summary:
Right after a mount is created, the inodes can't be accessed just yet as FSCK
or other setup might be running. Since periodic tasks might be running at the
same time as a mount is occuring, we need to make sure the mount is ready to
start a GC. The function EdenServer::getMountPoints does exactly that.

Reviewed By: chadaustin

Differential Revision: D42088584

fbshipit-source-id: 1fecb5c660246ddb9b25ce39d83990ee93907052