bssm: Support querying suffixes
Summary:
The BSSM optimisation is currently just used for queries to `commit_find_files` that specify basenames, not suffixes. That is because the logic for suffixes was added after I had already added but not yet landed BSSM.

This diff makes BSSM be used for that optimisation as well, as was intended for BSSM.

To do that, we traverse the sharded map to find all (key, value) pairs in which `key` has the given prefix (and that's why we store the basenames reversed). The logic for fetching keys by prefix is added on D42071060. From there we continue the search as normal.

The order of files will not be lexicographical anymore, but will be consistent between requests, so it can still be used to page requests.

Reviewed By: mitrandir77

Differential Revision: D42098314

fbshipit-source-id: 45f16be358e62be09c2542cb53e5af01a687af8e