reviewstack: add a fallback path when a GHE avatar cannot load
Summary:
As noted in the comments in the code, we cannot always load
the image for a GHE avatar because the browser may block it with CORB.
Some suggest using a proxy to fetch the avatar, but that is a non-starter
for ReviewStack. Instead, we introduce a rudimentary fallback
mechanism that uses the first letter of the username as an avatar.

Reviewed By: quark-zju

Differential Revision: D42158617

fbshipit-source-id: bf0eb63d190820c58de17cf4c3578b09bdbf65f6