reviewstack: support GitHub Enterprise users in login flow
Summary:
This makes it possible to use reviewstack.dev with a
GitHub Enterprise account. Unfortunately, it does not
appear that our existing Netlify OAuth flow can be used for
GHE accounts. I believe for that to work, the GHE
admin would have to create a separate instance of
ReviewStack with its own OAuth secret and hosted
on a different domain.

Fortunately, GHE users can still try reviewstack.dev:
it's just that they have to enter a little more information manually.
It turns out that if you have already bothered to authenticate
with your GHE account via the `gh` CLI, it is straightforward
to dump the PAT you need to try out ReviewStack.

Fixes https://github.com/facebook/sapling/issues/188

Reviewed By: quark-zju

Differential Revision: D42161860

fbshipit-source-id: 2982592152d46a8bc63267489a68e36353c0d24f