reviewstack: developer dialog support for GHE
Summary:
Amends the developer dialog page (i.e., the one you get when
running `yarn start` locally, which is different than what is in
prod, which is `NetlifyLoginDialog.tsx`) to make it possible
to specify the hostname for the GitHub instance you want to
work with, which defaults to github.com.

This value is stored in `localStorage` as the value
associated with `"github.hostname"`.

Note it is already the case that `localStorage.clear()` is called
when the user logs out, so we don't have to worry about that
case.

The trickier thing is ensuring that the property is persisted
correctly in the first place because we also call `localStorage.clear()`
before storing the user's PAT upon login.

This diff extends `gitHubTokenPersistence` to get things working
for now, though this is a bit racy since `github.hostname` is written
first and then `github.token` is written afterwards in the callback
to a `Promise`. Ideally, these values would be written atomically,
but that requires changing the contract of `gitHubTokenPersistence`,
which will be done in a subsequent diff.

Reviewed By: quark-zju

Differential Revision: D42156925

fbshipit-source-id: a008f948c836d94703fd0cb226ba2aea545c6b32