watchman: add metrics for Rust watchman
Summary:
Log the number of files returned and whether it was a fresh instance. Also, instrument the watchman query so it shows up in perftrace, and populates the watchmanquery_time field.

Note that I didn't go through the hgmetrics stuff and instead logged straight to tracing. hgmetrics currently only makes it to scuba for Python commands.

Reviewed By: quark-zju

Differential Revision: D42088975

fbshipit-source-id: 38a2a647f9e3ce9b612d8f203fc3777c259698b9