run ghstack from the current commit instead of the branch tip
Summary:
Previously, in our Sapling port of ghstack, we were interpreting `HEAD`
from Git as the result of:

```
sl log -r 'max(descendants(.))' -T '{node}'
```

which was forcing ghstack to always operate from the tip of your feature
branch rather than where you are. This updates the code so that `HEAD`
is interpreted as `sl whereami` so you can submit only your current commit
and the commits *below* it in the stack as pull requests.

Fixes https://github.com/facebook/sapling/issues/283.

Reviewed By: muirdm

Differential Revision: D42121060

fbshipit-source-id: 9ea54c200b9d7938225a858ef2421bc5cf4b78e5