implement `sl pr list`
Summary:
GitHub's `gh` CLI has a number of useful subcommands under `gh pr`.
When run in a Git repo, `gh` can automatically figure out the appropriate
value to use with the `--repo` flag, though that is not the case in Sapling.

In this case, we do a small amount of plumbing to implement
`sl pr list [flags]` so that it execs the analogous `gh pr list [flags]` call.
This pattern should work well for supporting more `gh pr` commands.

Reviewed By: muirdm

Differential Revision: D42120917

fbshipit-source-id: 51bd51f46cd15fc9e4bbde631adfe24646b0efb7