pr: enable prmarker for all git repos
Summary:
Currently we enabled `prmarker` extension in sapling.rs config, and also added a post pull hook for calling the command of `prmarker` extension, but in some corner cases, the hook triggers other `sl` binary, and will cause "no debugprmarker" warning message.

```
post-pull.prmarker=sl debugprmarker

```

This diff is to enable `prmarker` for all git repos. BTW: it's safe to enable this extension for non-GitHub repo, it will just exit if the repo is not a GitHub repo.

Reviewed By: quark-zju

Differential Revision: D42181509

fbshipit-source-id: 976009a570bb476194f69b99feabf7256467f320