remotenames: translate 'push --to remote/foo' to 'push default --to foo'
Summary:
The command flags of `push` is not obvious.  A common mistake is
`push --to remote/main` (which will create a `remote/remote/main` remote
bookmark).

This diff changes the logic to accept `push remote/main`,
`push --to remote/main` or `push default/main` to do the obvious thing for a
better UX.

Reviewed By: muirdm

Differential Revision: D42203177

fbshipit-source-id: d73647823c147dbdbce58c71a04dd4965e8ac2fb