temporary fix to website command codegen to work around escaping issue (#352)
Summary:
We currently have an issue where the generated docs for
`sl pr` produce a line that happens to start with the word `import`.
MDX tries to parse this line as an import statement.
As a quick fix so we can publish the website again,
introduce a simple escaping solution as described on:

https://stackoverflow.com/questions/61188612/escaping-import-and-react-html-to-display-as-text-in-mdx/74096776

Pull Request resolved: https://github.com/facebook/sapling/pull/352

Test Plan:
Imported from OSS

Verified in D42230100

Reviewed By: zzl0

Differential Revision: D42230099

Pulled By: bolinfest

fbshipit-source-id: af2f87e40ed7d71dac5eba7a054f026a1209ec62