tests: deflake test-clone-resume.t
Summary:
Previously it triggered a checkout error by setting bad permissions on a directory checkout needed to write to. However, nothing guaranteed that other files were successfully written to before the permission error, so sometimes zero files were skipped when continuing the checkout (which failed the test). Now we use a pre-existing failpoint to fail the checkout. Hopefully this works now on Windos as well.

I also reduced the size of test data from 5002 files to 2.

Reviewed By: zzl0

Differential Revision: D42256394

fbshipit-source-id: d3016779b5bbd734a0f00afd2885b63218dedaae