notifications: properly handle exception when creating E-Menu
Summary:
The EdenFS telemetry shows that 90% the vast majority of all the EdenFS crashes
on Windows occur very shortly after starting up, with most of these pointing to
~thread failing. This is triggered by a non-detached thread not being joined.
Looking at backtraces containing symbols and sources, the crash appears to
occur due to the windowsEventLoop throwing an exception but never being joined,
due to the ~WindowsNotifier never being called.

To fix this, we can simply split the WindowsNotifier constructor in 2: the
construtor, and an initialize method that may throw. This ensures the
destructor is always called.

Reviewed By: MichaelCuevas

Differential Revision: D42272618

fbshipit-source-id: bf092573b3e4de2210a2b9c6124bf93c718cb3ae