commit_graph: rename test source file and remove conditional compilation
Summary:
Fix OSS getdeps build of Mononoke, which has been broken by default cargo testing behaviour.

The tests must not be in a directory named `tests`.  If they are, Cargo helpfully detects them and tries to compile them, but it does that in the context of the parent directory's `Cargo.toml`, which has the wrong dependencies.  Our pattern of putting crates inside other crates is potentially harmful for Cargo behaviours like this, so it's better to split these out to real sibling crates.

As such, move `testlib` to a sibling crate (`commit_graph_testlib`).  In order to avoid a circular dependency between `testlib` and the main `commit_graph` crate, move the in-memory storage implementation and the corresponding tests inside the `testlib` crate, rather than having them in a separate binary.  This is overall simpler, and works ok with the generated `Cargo.toml` files.

Reviewed By: YousefSalama

Differential Revision: D42280242

fbshipit-source-id: 2a2d2721755097a2e498706443a1513eea195503