introduce ThriftStreamStartupStatusSubscriber
Summary:
In D38962809 I am teaching the eden CLI to show eden startup status as start is
happening.

In this change we are creating the startup subscriber for a thrift stream.

This is a wrapper class that handles creating a thrift stream and registering
it with the startup status state.

Nicely handling the completed/cancel callback is non trivial and I made a
simplifying decision here that keeps around memory for longer. Inspect
carefully!

Reviewed By: xavierd

Differential Revision: D41057242

fbshipit-source-id: 5760eece2bc8843d4d5ed047f9aeab2d3922d453