telemetry: enable SubprocessScribeLogger on Windows
Summary:
I was looking at adding telemetry for when the E-Menu fails to start in
D42272618 (https://github.com/facebook/sapling/commit/eda170e8b49ab6d280d4470823475d717af154fd) but remembered that in-process telemetry wasn't enabled on Windows.
Attempting to make it compile succeeded and with a trivial tweak was even able
to get telemetry!

My understanding as to why this "just works" is probably historical: this code
uses to rely on folly::Subprocess which only runs on Linux and macOS as it
relies on fork+exec. However, the code was switched to use SpawnProcess which
also runs on Windows, and thus removing the ifdef was sufficient.

Reviewed By: chadaustin

Differential Revision: D42273007

fbshipit-source-id: 8377328599349da0687414314ea13d2a54bbce94