blobrepo: remove get_generation_number
Summary:
Make `BlobRepo` like an ordinary facet container by removing its specialized methods.

In this diff, we replace the `get_generation_number` method with the equivalent method (`get_generation_number`) on the changeset fetcher.  The fetcher version doesn't return `Option<Generation>`, but it seems nobody was using that feature of the old method anyway, and it's clearer if things are consistent.

Reviewed By: mitrandir77

Differential Revision: D42314995

fbshipit-source-id: b1be08020d7589869a6c7ad9077fd87b17aec271