commit_graph: implement add_recursive method that adds a changeset and its missing ancestors
Summary: When migrating from the current changeset facet to the new commit graph facet, there will be a time after backfilling the new commit graph when it's not fully in sync with the changeset facet, and to insert a changeset to the commit graph we need all its parents to be already inserted. This implements a method that recursively tries to insert ancestors of a changeset from the changeset facet to the commit graph to handle this case.

Reviewed By: markbt

Differential Revision: D42274344

fbshipit-source-id: a0776af7f74b20e1fdd30cb83d722fd3c7a5963e