commit_graph: Optimising backfiller
Summary:
yancouto: Before this diff, the backfiller was doing about 2 commit a second. With this one, it does about 2k, 1000x faster :D

The problem before was that we needed to do serial writes and reads to SQL for each commit. This is pretty bad.

The solution is to first do everything in memory, then "dump" it very quickly and in batch to SQL. The in-memory stuff is still serial, but it's order or magnitudes faster than going to SQL.

It's a bit tricky to write the edges in batch to SQL, because we depend on their ids, but I managed to do it by breaking into ~4 steps:
1. Write only the changesets to the table, without edges, so they get an id.
2. Read the ids
3. Write the edges now that we know the ids
4. Write the merge parents (THIS IS MISSING FROM THIS DIFF)

Reviewed By: liubov-dmitrieva

Differential Revision: D42227090

fbshipit-source-id: eb1589f32d4b9082064b83f09fee3ce5f3dfadd0