commit_graph_compat: implement ChangesetsCommitGraphCompat and use it in RepoFactory
Summary: Implements a compatibility layer between the Changesets trait and CommitGraph. For now it forwards all the reads to the old Changesets trait, but double writes to both Changesets and CommitGraph if a per repo tunable is set to true. Writes to CommitGraph have a timeout that is controlled by a tunable, they are also timed and logged to the scuba table mononoke_commit_graph, along with whether the write succeeded, timed out, or failed (and the error in this case).

Reviewed By: markbt

Differential Revision: D42295681

fbshipit-source-id: 6d929ff580141561accfd04342fabae8b28bdf9b