commit_graph: implement BufferedCommitGraphStorage to make backfiller resumable and fix a bug in FixEdges
Summary:
This adds a new CommitGraphStorage implementation called BufferedCommitGraphStorage, that has an in memory storage buffer that it adds new changeset edges to before flushing them to a persistent backend. By switching the backfiller to use this new storage, we don't have to store the entire commit graph in memory and can pause and resume backfilling, while still having the benefit of not writing serially to the database.

This also adds inserting merge parents to the add_many method that was missing in the previous diff, and fixes a bug in the FixEdges query.

Reviewed By: markbt

Differential Revision: D42267271

fbshipit-source-id: 2437480c9283392d52eef38cf40a55fd8618cb96